Cleanified structure/wscripts - added namespaces
