+++ hexapod_controller/README.md
@@ -14,7 +14,7 @@ A simple sinusoidal controller.
 
 ### Compile and install
 
-- cd to `hexapod_common` folder
+- cd to `hexapod_controller` folder
 - Configure with `./waf configure --prefix=path_to_install`
 - Compile with `./waf build`
 - Install with `./waf install`
@@ -43,29 +43,29 @@ def options(opt):
 
 @conf
 def check_hexapod_controller(conf):
-	includes_check = ['/usr/local/include/hexapod_controller', '/usr/include/hexapod_controller']
+	includes_check = ['/usr/local/include', '/usr/include']
 	libs_check = ['/usr/local/lib', '/usr/lib']
 
 	# You can customize where you want to check
 	# e.g. here we search also in a folder defined by an environmental variable
 	if 'RESIBOTS_DIR' in os.environ:
-		includes_check = [os.environ['RESIBOTS_DIR'] + '/include/hexapod_controller'] + includes_check
+		includes_check = [os.environ['RESIBOTS_DIR'] + '/include'] + includes_check
 		libs_check = [os.environ['RESIBOTS_DIR'] + '/lib'] + libs_check
 
 	if conf.options.controller:
-		includes_check = [conf.options.controller + '/include/hexapod_controller']
+		includes_check = [conf.options.controller + '/include']
 		libs_check = [conf.options.controller + '/lib']
 
 	try:
 		conf.start_msg('Checking for hexapod_controller includes')
-		res = conf.find_file('hexapod_controller_simple.hpp', includes_check)
+		res = conf.find_file('hexapod_controller/hexapod_controller_simple.hpp', includes_check)
 		conf.end_msg('ok')
 		conf.start_msg('Checking for hexapod_controller libs')
 		res = res and conf.find_file('libhexapod_controller_simple.a', libs_check)
 		conf.end_msg('ok')
 		conf.env.INCLUDES_HEXAPOD_CONTROLLER = includes_check
-		conf.env.LIBPATH_HEXAPOD_CONTROLLER = libs_check
-		conf.env.LIB_HEXAPOD_CONTROLLER = ['hexapod_controller_simple']
+		conf.env.STLIBPATH_HEXAPOD_CONTROLLER = libs_check
+		conf.env.STLIB_HEXAPOD_CONTROLLER = ['hexapod_controller_simple']
 	except:
 		conf.end_msg('Not found', 'RED')
 		return
@@ -76,11 +76,11 @@ Then in your C++ code you would have something like the following:
 
 ```cpp
 // previous includes
-#include <hexapod_controller_simple.hpp>
+#include <hexapod_controller/hexapod_controller_simple.hpp>
 
 // rest of code
 
-HexapodControllerSimple controller(controller_parameters, broken_legs);
+hexapod_controller::HexapodControllerSimple controller(controller_parameters, broken_legs);
 
 // rest of code
 ```
