+++ hexapod_controller/wscript
@@ -39,10 +39,9 @@ def configure(conf):
 
 def build(bld):
     bld.stlib(features = 'cxx cxxstlib',
-              source=' \
-              hexapod_controller_simple.cpp',
-              includes = '.',
+              source='src/hexapod_controller_simple.cpp',
+              includes = './include',
               target='hexapod_controller_simple')
 
-    bld.install_files('${PREFIX}/include/hexapod_controller', 'hexapod_controller_simple.hpp')
+    bld.install_files('${PREFIX}/include/hexapod_controller', 'include/hexapod_controller/hexapod_controller_simple.hpp')
     bld.install_files('${PREFIX}/lib', 'build/libhexapod_controller_simple.a')
